[BAR-685][AK] move health check file under folder
